Fix failing unit spec
The spec "should return a ResourceError exception if remote URI is
invalid" expects the URI "http://example.com/file.tgz" to be invalid and
to throw a ResourceError exception but on my machine that URI is valid,
returns a 404 code, and throws a ResourceNotFound error. Fixed to use a
URI that's actually invalid.